feat: Add `PopoverDisclosureArrow` component
